Shift core plugin ApplicationMailer/BaseMailer stuff around to be more consistent with feature plugins
